+++ app/src/main/java/com/example/alldayfit/count/model/ExerciseRecord.kt
@@ -1,4 +1,4 @@
-package com.example.alldayfit.count
+package com.example.alldayfit.count.model
 
 data class ExerciseRecord(
     val timestamp: String
