--- app/src/main/res/drawable/progress_circular.xml
@@ -1,20 +1,0 @@
-<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:id="@android:id/background">
-        <shape
-            android:innerRadiusRatio="2.5"
-            android:shape="ring"
-            android:thicknessRatio="15.0"
-        android:useLevel="false">
-        <solid android:color="@color/gray" />
-    </shape>
-</item>
-<item android:id="@android:id/progress">
-<shape
-    android:innerRadiusRatio="2.5"
-    android:shape="ring"
-    android:thicknessRatio="15.0"
-android:useLevel="true">
-<solid android:color="@color/orange" />
-</shape>
-    </item>
-    </layer-list>
