--- app/src/main/res/layout/activity_sign_in_page.xml
@@ -1,91 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".settings.SignInPage">
-
-    <TextView
-        android:id="@+id/logo"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="All-Day Fit"
-        android:textSize="48sp"
-        android:layout_marginTop="50dp"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        />
-
-    <ImageView
-        android:id="@+id/btn_naver"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:src="@drawable/btn_naver"
-        android:layout_marginHorizontal="40dp"
-        android:layout_marginTop="550dp"
-        app:layout_constraintDimensionRatio="7"
-        app:layout_constraintTop_toBottomOf="@id/logo"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintBottom_toTopOf="@id/btn_google"/>
-
-    <ImageView
-        android:id="@+id/ic_naver"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:src="@drawable/ic_naver"
-        android:layout_marginStart="7dp"
-        app:layout_constraintDimensionRatio="1"
-        app:layout_constraintTop_toTopOf="@id/btn_naver"
-        app:layout_constraintStart_toStartOf="@id/btn_naver"
-        app:layout_constraintBottom_toBottomOf="@id/btn_naver"
-        app:layout_constraintEnd_toStartOf="@id/ic_naver"/>
-
-    <TextView
-        android:id="@+id/naver_login"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="네이버 로그인"
-        android:layout_marginStart="5dp"
-        app:layout_constraintTop_toTopOf="@id/btn_naver"
-        app:layout_constraintBottom_toBottomOf="@id/btn_naver"
-        app:layout_constraintStart_toEndOf="@id/ic_naver"/>
-
-    <ImageView
-        android:id="@+id/btn_google"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:src="@drawable/btn_google"
-        android:layout_marginBottom="30dp"
-        app:layout_constraintDimensionRatio="7"
-        android:elevation="20dp"
-        android:layout_marginHorizontal="30dp"
-        app:layout_constraintTop_toBottomOf="@id/btn_naver"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"/>
-
-    <ImageView
-        android:id="@+id/ic_google"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:src="@drawable/ic_google"
-        android:elevation="20dp"
-        android:layout_marginStart="16dp"
-        app:layout_constraintStart_toStartOf="@id/btn_google"
-        app:layout_constraintTop_toTopOf="@id/btn_google"
-        app:layout_constraintBottom_toBottomOf="@id/btn_google"/>
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="구글 로그인"
-        android:layout_marginHorizontal="8dp"
-        android:layout_marginStart="20dp"
-        android:elevation="20dp"
-        app:layout_constraintTop_toTopOf="@id/btn_google"
-        app:layout_constraintBottom_toBottomOf="@id/btn_google"
-        app:layout_constraintStart_toEndOf="@id/ic_google"/>
-</androidx.constraintlayout.widget.ConstraintLayout>
