+++ app/src/main/res/layout/count_item.xml
@@ -1,44 +1,31 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
-    android:layout_height="wrap_content">
-
-    <View
-        android:id="@+id/item_back"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:layout_marginBottom="10dp"
-        android:elevation="30dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        android:background="@drawable/item_radius"/>
+    android:layout_height="60dp"
+    android:layout_margin="5dp"
+    android:background="@drawable/item_radius"
+    android:elevation="30dp">
 
     <TextView
         android:id="@+id/exercise_titele"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginTop="25dp"
-        android:layout_marginBottom="25dp"
-        android:elevation="30dp"
+        android:layout_marginStart="35dp"
         android:text="렛 풀 다운"
-        app:layout_constraintBottom_toBottomOf="@+id/item_back"
-        app:layout_constraintEnd_toEndOf="@+id/item_back"
-        app:layout_constraintStart_toStartOf="@+id/item_back"
-        app:layout_constraintTop_toTopOf="@+id/item_back" />
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
         android:id="@+id/set"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginEnd="10dp"
-        android:layout_marginBottom="10dp"
+        android:layout_marginEnd="25dp"
         android:elevation="30dp"
-        android:text=""
-        app:layout_constraintBottom_toBottomOf="@+id/item_back"
-        app:layout_constraintEnd_toEndOf="@+id/item_back" />
+        android:text="1시간/3세트 30회"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
