+++ app/src/main/res/values/styles.xml
@@ -56,10 +56,11 @@
     <style name="buttonText">
         <item name="android:textSize">@dimen/text_size_small</item>
     </style>
-
+      
     <style name="CircularProgressBar">
+        <item name="android:radius">10dp</item>
         <item name="android:indeterminateOnly">false</item>
-        <item name="android:progressDrawable">@drawable/progress_circular</item>
+        <item name="android:progressDrawable">@drawable/progress_gradation</item>
     </style>
 
 </resources>
